Move Tags component to common dir
