+++ assets/js/components/SapSystemsOverview/SapSystemsOverview.jsx
@@ -9,7 +9,7 @@ import PageHeader from '@components/PageHeader';
 import HealthIcon from '@components/Health';
 import Table from '@common/Table';
 import SAPSystemItemOverview from '@components/SapSystemsOverview/SapSystemItemOverview';
-import Tags from '@components/Tags';
+import Tags from '@common/Tags';
 import HealthSummary from '@components/HealthSummary/HealthSummary';
 import DeregistrationModal from '@components/DeregistrationModal';
 import { getCounters } from '@components/HealthSummary/summarySelection';
